StatsCollectionSubphase
